split component for chart
